refactor!: qualify the impl-package exports and move SqlLog's rendering internals out of the public class
The st.orm.core.template.impl and st.orm.core.repository.impl packages are
exported to Storm's own modules only, so the module descriptor states and
enforces that their types are not API. SqlLog keeps the Scope/Recorder
diagnostics API; summary rendering and hydration-shape analysis move to
SqlLogRenderer and call-site capture to CallSiteCapture, together with the
display setters, since how summaries render is configured through the
storm.sql_log.* properties and the Spring and Ktor keys rather than
programmed.

Fixes #395
